DPLL with advanced simplification
This version 2.0 DPLL algorithm implements advanced simplification method. This method constructs a binary searching tree with current results in an array. Without genuinely copying and simplifying a large dataset as previously we did, we shrink the time complexity of simplificaion process to O(1) level, and also massively lower the RAM cost.